+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Release 5.11.0 (unreleased):
  - OpenID for Verifiable Presentations:
    - In `RequestOptions` deprecate property `encryption`, as this depends on the response mode
    - In `AuthnResponseResult` returned from `OpenId4VpVerifier.validateAuthnResponse()` remove parameter `state`
+   - In `OpenId4VpVerifier` remove `validateAuthnResponse(input: Map)`
+   - In `OpenId4VpVerifier` add option to provide `externalId` when validating authn responses, useful for DCAPI flows
 
 Release 5.10.0:
  - StatusListToken: